fix/ meetperry add additional error handling

+++ managers/webhook.py
@@ -1,8 +1,10 @@
 import logging
+from datetime import datetime
 
 import psycopg2
 
-from exceptions.webhook import InvalidEventStatusException, DuplicateTaskCreationException, TaskNotFoundException
+from exceptions.webhook import InvalidEventStatusException, DuplicateTaskCreationException, TaskNotFoundException, \
+    TaskInvalidStatusException
 from managers.data import DataManager
 from schemas.webhook import WebhookPayload
 
@@ -10,15 +12,23 @@ from schemas.webhook import WebhookPayload
 class WebhookManager(DataManager):
 
     async def processWebhook(self, payload: WebhookPayload):
+        task_exists = await self.taskExists(payload)
+        completed_date, deleted_date = await self.getTaskStatus(payload)
         if payload.event == "created":
             await self.createTask(payload)
         elif payload.event == "updated":
-            if not await self.taskExists(payload):
+            if not task_exists:
                 raise TaskNotFoundException()
+            elif completed_date is not None:
+                return
+            elif deleted_date is not None:
+                raise TaskInvalidStatusException()
             await self.updateTask(payload)
         elif payload.event == "deleted":
             if not await self.taskExists(payload):
                 raise TaskNotFoundException()
+            if deleted_date is not None:
+                return
             await self.deleteTask(payload)
         else:
             raise InvalidEventStatusException()
@@ -54,3 +64,13 @@ class WebhookManager(DataManager):
         sql = "SELECT EXISTS(SELECT 1 FROM meetperry.task WHERE external_id = %s);"
         data = [payload.metadata.id]
         return self.postgres.exists(sql, data)
+
+    async def getTaskStatus(self, payload: WebhookPayload) -> (datetime, datetime):
+        logging.debug(f"Retrieving task for id {payload.metadata.id}")
+        sql = "SELECT completed_date, deleted_date FROM meetperry.task WHERE external_id = %s;"
+        data = [payload.metadata.id]
+        row = self.postgres.execute_one(sql, data)
+        if row is None:
+            return None, None
+        else:
+            return row[0], row[1]
